Update README.md
The URL Shortener is a simple web application built with Python using the Flask framework. It provides a convenient way to shorten long URLs, making them more manageable and shareable. The program allows users to enter a long URL through a web form and generates a unique shortened URL for the input.